+++ packages/coin-cro/src/config/cosmos.proto
