+++ packages/coin-cro/src/config/messages.js
