+++ packages/coin-cro/src/config/params.ts
@@ -1,45 +1,11 @@
-export const COIN_TYPE = '8000003c';
+export const COIN_TYPE = '8000018a';
 
 export const TRANSFER = {
-  script: `03040601C707000000003CA00700C2ACD70032FFF8C2ACD7001EFFF6C2ACD70028FFF6CC071094CAA02700C2A2D700FFF6CC071080CC0E1019C2E09700CC07C0028080BE0710DC07C00343524FCC0FC0023078BAA02F6C0E04DDF09700DAA2D7C0FFF612D207CC05065052455353425554546F4E`,
-  signature: `003045022100B3D3C5F089BFD31021357B05F769256299460DAA54B4DBD8777158C047A3729E02206078F1403E17FB750E7B4BAE197566165AC5F5061C0764F325AC7327EC2B5EB9`,
-  get scriptWithSignature(): string {
-    return this.script + this.signature;
-  },
-};
+    script: '03030201C707000000018ACC07100aA00700CC07100aA00700CC07C01E0a1c2f636f736d6f732e62616e6b2e763162657461312e4d736753656e64CC071012A00700CC07100aBFACD7C021FFC002CC071012BFACD7C061FFC002CC07101aA00700CC07C0090a076261736563726fCC071012A00700BAACC76CA1080D01BE0700BE0700BE0700BE0700CC071012A0070029AC9700C9CAAC9700C9BE0700BE0700CC071012A00700CC07100aA00700CC07C0270a460a1f2f636f736d6f732e63727970746f2e736563703235366b312e5075624b657912230a21CAA0C70021CC07C00612040a020801CC071018BFACC700C108BE0700CC071012A00700CC07100aA00700CC07C0090a076261736563726fCC071012A00700BAACC76CA9080D01BE0700BE0700CC071010BFACC700B108BE0700BE0700CC07C01C1a1A63727970746f2d6f72672d636861696e2d6d61696e6e65742d31CC071020BFACC700B908DC07C00343524FDDACD70061FFC0DAACC7C0A10808D207CC05065052455353425554546F4E',
+    signature: '304502206FD556F37F2390835CA56A91CCB9E7005FC56C74551FEB8CB270BF835C1B7B1B0221008838C3BE0E2237ED3C7226418D08802501337A1078142E11D4FD8198BDF2B5C8'
+}
 
-export const ERC20 = {
-  script: `03040601C707000000003C11ACC7CC3C0F044B1507C002FF00A00700C2ACD70034FFF8C2A5D700FFF6C2ACD7002AFFF6CC071094CAAC27009CCC07C01380B844a9059cbb000000000000000000000000CAA02700CC07200000000000000000000000000000000000000000CAA2C7000C12AC17C03C0401061507C002FF00B5AC17003CC2ACB7003DCC07C0028080BE071012AC17C0430401071507C002FF00B5AC170043CAACBE0044DEE09700250E0011ACC7CC931D04B01507C004CC0F104012AC17C0940401071507C002FF00B5AC170094CAACBF0095DEF09700250F00CC0FC0023078BAA02F6C0E04DDF0970012AC17C0930400141507C002FF00B5AC170093DAA2C7B00CD207CC05065052455353425554546F4E`,
-  signature:
-    `3044022009285c34eee70211b248b9efeb4881e70ba57e1c261e3b8144bc145014a35b8c022046d03a39667ab4654769bc2e81e70c0b4ec4ed914717d00f6ef650c357281028`.padStart(
-      144,
-      '0'
-    ),
-  get scriptWithSignature(): string {
-    return this.script + this.signature;
-  },
-};
+export enum TX_TYPE_URL{
+    SEND = '/cosmos.bank.v1beta1.MsgSend',
+}
 
-export const SmartContract = {
-  script: `03040601C707000000003CA00700C2ACD70032FFF8C2ACD7001EFFF6C2ACD70028FFF6CC071094CAA02700C2A2D700FFF6C2AC97003ACC0E1019C2E09700CC07C0028080BE0710DC07C00343524FD207C005534D415254D207CC05065052455353425554546F4E`,
-  signature:
-    `3045022072902FD446A0E80824EAEF45ACCA5F800BE52AD3CE9D011AF90C8D85710059D0022100C467A0FE1BF941B4900CB269B80BD2D1999EEB5B8633676AA0383E8EDE863FA3`.padStart(
-      144,
-      '0'
-    ),
-  get scriptWithSignature(): string {
-    return this.script + this.signature;
-  },
-};
-
-export const SmartContractSegment = {
-  script: `03050601C707000000003CA00700C2ACD70032FFF8C2ACD7001EFFF6C2ACD70028FFF6CC071094CAA02700C2A2D700FFF6C4ACC7003A04CC0E1019C2E09700CC07C0028080BE0710DC07C00343524FD207C005534D415254D207CC05065052455353425554546F4E`,
-  signature:
-    `3046022100A0188B33A3CCCF98BF055FB82BF621E9D43AFE7DD066D907DD446688E91141C0022100EF67E474862FA59DC5DE49079FD5258CDB6939C41C84E0170132399E1CB96592`.padStart(
-      144,
-      '0'
-    ),
-  get scriptWithSignature(): string {
-    return this.script + this.signature;
-  },
-};
