+++ packages/coin-cro/src/utils/scriptUtil.ts
