+++ packages/coin-cronos/.babelrc
