+++ packages/coin-cronos/package.json
@@ -1,15 +1,25 @@
 {
-  "name": "@coolwallet/cro",
+  "name": "@coolwallet/cronos",
   "version": "1.0.0",
-  "description": "Coolwallet Crypto.org Chain sdk",
+  "description": "Coolwallet Cronos sdk",
   "main": "lib/index.js",
   "author": "coolwallet-team",
   "license": "Apache-2.0",
   "scripts": {
     "build:types": "tsc --emitDeclarationOnly",
-    "build:ts": "babel src --out-dir lib --extensions \".ts,.tsx,.js\" --source-maps inline",
-    "build": "npm run build:types && npm run build:ts",
-    "test": "jest --verbose"
+    "build:ts": "babel src --out-dir lib --extensions \".ts,.tsx\" --source-maps inline",
+    "build": "del-cli lib && npm run build:types && npm run build:ts",
+    "test": "jest"
+  },
+  "dependencies": {
+    "bip32": "^2.0.4",
+    "eth-sig-util": "^3.0.1",
+    "keccak": "^3.0.2",
+    "rlp": "^2.2.7",
+    "web3": "^1.7.0"
+  },
+  "peerDependencies": {
+    "@coolwallet/core": "^1.0.14"
   },
   "devDependencies": {
     "@babel/cli": "^7.16.8",
@@ -17,9 +27,11 @@
     "@babel/plugin-proposal-class-properties": "^7.7.4",
     "@babel/preset-env": "^7.7.1",
     "@babel/preset-typescript": "^7.7.7",
-    "@coolwallet/core": "file:../core",
+    "@coolwallet/core": "^1.0.14",
     "@types/eth-sig-util": "^2.1.1",
     "@types/node": "^14.0.13",
+    "@types/keccak": "^3.0.1",
+    "del-cli": "^4.0.1",
     "typescript": "^4.4.3"
   },
   "repository": {
@@ -28,20 +40,5 @@
   },
   "bugs": {
     "url": "https://github.com/CoolBitX-Technology/coolwallet3-sdk/issues"
-  },
-  "dependencies": {
-    "@cosmostation/cosmosjs": "^0.11.1",
-    "@types/keccak": "^3.0.1",
-    "bech32": "^2.0.0",
-    "bip32": "^2.0.4",
-    "eth-sig-util": "^3.0.1",
-    "keccak": "^3.0.2",
-    "protobufjs": "^6.11.2",
-    "protocol-buffers": "^4.2.0",
-    "protocol-buffers-encodings": "^1.1.1",
-    "rlp": "^2.2.7",
-    "varint": "^6.0.0",
-    "varstruct": "^6.1.3",
-    "web3": "^1.7.0"
   }
 }
