+++ packages/coin-cronos/src/config/chain.ts
