+++ packages/coin-cronos/src/config/tokenType.ts
