+++ packages/coin-cronos/src/utils/scriptUtils.ts
