+++ packages/coin-cronos/src/utils/stringUtil.ts
