--- packages/coin-org/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    /* Basic Options */
-    "target": "esnext", /* Specify ECMAScript target version: 'ES3' (default), 'ES5', 'ES2015', 'ES2016', 'ES2017', 'ES2018', 'ES2019' or 'ESNEXT'. */
-    "module": "commonjs", /* Specify module code generation: 'none', 'commonjs', 'amd', 'system', 'umd', 'es2015', or 'ESNext'. */
-    "declaration": true, /* Generates corresponding '.d.ts' file. */
-    "outDir": "lib", /* Redirect output structure to the directory. */
-    /* Strict Type-Checking Options */
-    "strict": true, /* Enable all strict type-checking options. */
-    "allowSyntheticDefaultImports": true, /* Allow default imports from modules with no default export. This does not affect code emit, just typechecking. */
-    "esModuleInterop": true, /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
-    "skipLibCheck": true
-  },
-  "typedocOptions": {
-    "exclude": [
-      "src/declarations.ts"
-    ]
-  }
-}
